Conflicts:
	Makefile
	bauer/fabrikbauer.cc
	bauer/hausbauer.h
	bauer/tunnelbauer.cc
	bauer/wegbauer.cc
	besch/writer/tunnel_writer.cc
	besch/writer/vehicle_writer.cc
	dataobj/einstellungen.cc
	dataobj/einstellungen.h
	dataobj/fahrplan.cc
	dataobj/network.cc
	dataobj/network_cmd_ingame.cc
	dataobj/network_file_transfer.cc
	dings/baum.cc
	dings/bruecke.cc
	dings/leitung2.cc
	dings/leitung2.h
	freight_list_sorter.cc
	gui/climates.cc
	gui/components/gui_label.h
	gui/components/gui_numberinput.cc
	gui/components/gui_scrollpane.h
	gui/components/gui_textinput.cc
	gui/convoi_detail_t.cc
	gui/convoi_info_t.cc
	gui/depot_frame.cc
	gui/depot_frame.h
	gui/display_settings.cc
	gui/enlarge_map_frame_t.cc
	gui/goods_frame_t.cc
	gui/line_management_gui.cc
	gui/loadsave_frame.cc
	gui/money_frame.cc
	gui/savegame_frame.cc
	gui/settings_stats.cc
	makeobj/makeobj.cc
	player/ai.cc
	player/ai.h
	player/ai_goods.cc
	player/ai_passenger.cc
	player/simplay.cc
	player/simplay.h
	simcity.cc
	simconvoi.cc
	simdepot.cc
	simdepot.h
	simfab.cc
	simfab.h
	simgraph16.cc
	simhalt.cc
	simhalt.h
	simline.cc
	simmain.cc
	simutrans/config/simuconf.tab
	simversion.h
	simware.cc
	simware.h
	simwerkz.cc
	simworld.cc
	simworld.h
	tpl/vector_tpl.h
	vehicle/simvehikel.cc
	vehicle/simvehikel.h

+++ dataobj/dingliste.h
@@ -20,7 +20,7 @@ private:
 	*/
 	uint8 top;
 
-	void set_capacity(uint8 new_cap);
+	void set_capacity(uint16 new_cap);
 
 	bool grow_capacity();
 
@@ -74,7 +74,7 @@ public:
 	bool loesche_alle(spieler_t *sp,uint8 offset);
 	bool ist_da(const ding_t* obj) const;
 
-	inline int get_top() const {return top;}
+	inline uint8 get_top() const {return top;}
 
 	/**
 	 * sorts the trees according to their offsets
