+++ gui/gui_convoiinfo.cc
@@ -63,7 +63,7 @@ void gui_convoiinfo_t::zeichnen(koord offset)
 
 		int w = display_proportional_clip(pos.x+offset.x+2,pos.y+offset.y+6+LINESPACE, translator::translate("Gewinn"), ALIGN_LEFT, COL_BLACK, true)+2;
 		char buf[256];
-		money_to_string(buf, (double)(cnv->get_jahresgewinn()/100) );
+		money_to_string(buf, cnv->get_jahresgewinn() / 100.0 );
 		w += display_proportional_clip(pos.x+offset.x+2+w+5,pos.y+offset.y+6+LINESPACE, buf, ALIGN_LEFT, cnv->get_jahresgewinn()>0?MONEY_PLUS:MONEY_MINUS, true);
 		max_x = max(max_x,w+5);
